Move external to submodule.
