--- app/src/main/java/ayds/winchester/songinfo/moredetails/data/wikipedia/repository/external/wikipedia/WikipediaTrackService.kt
@@ -1,7 +1,0 @@
-package ayds.winchester.songinfo.moredetails.data.wikipedia.repository.external.wikipedia
-
-import ayds.winchester.songinfo.moredetails.domain.entity.Info.ArtistInfo
-
-interface WikipediaTrackService {
-    fun getInfo(artistName: String): ArtistInfo?
-}
