+++ app/src/main/java/ayds/winchester/songinfo/moredetails/data/wikipedia/repository/local/wikipedia/sqldb/CursorToWikipediaInfoMapper.kt
@@ -1,7 +1,7 @@
 package ayds.winchester.songinfo.moredetails.data.wikipedia.repository.local.wikipedia.sqldb
 
 import android.database.Cursor
-import ayds.winchester.songinfo.moredetails.domain.entity.Info.ArtistInfo
+import ayds.winchester2.wikipediaexternal.data.wikipedia.entity.Info.ArtistInfo
 import java.sql.SQLException
 
 interface CursorToWikipediaInfoMapper {
