--- app/src/main/java/ayds/winchester/songinfo/moredetails/domain/entity/Info.kt
@@ -1,12 +1,0 @@
-package ayds.winchester.songinfo.moredetails.domain.entity
-
-sealed class Info {
-
-    data class ArtistInfo(
-        var description: String,
-        var wikipediaURL: String,
-        var isLocallyStored: Boolean = false,
-    ): Info()
-
-    object EmptyInfo: Info()
-}
