+++ app/src/main/java/ayds/winchester/songinfo/moredetails/domain/repository/WikipediaRepository.kt
@@ -1,6 +1,6 @@
 package ayds.winchester.songinfo.moredetails.domain.repository
 
-import ayds.winchester.songinfo.moredetails.domain.entity.Info
+import ayds.winchester2.wikipediaexternal.data.wikipedia.entity.Info
 
 interface WikipediaRepository {
     fun getInfo(artist: String): Info
